Move all the lpa specific steps to a new context file.
